# docs/swagger.yaml
basePath: /
definitions:
  domain.Booking:
    properties:
      end_time:
        type: string
      id:
        type: integer
      service_id:
        type: string
      start_time:
        type: string
      status:
        description: pending, confirmed, completed
        type: string
      user_id:
        type: string
      vendor_id:
        type: integer
    type: object
  domain.VendorSummary:
    properties:
      statusCounts:
        additionalProperties:
          type: integer
        type: object
      totalBookings:
        type: integer
      vendorID:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Swagger API for the BRB Mid Service Platform.
  title: BRB Mid SVC Platform API
  version: "1.0"
paths:
  /bookings:
    post:
      consumes:
      - application/json
      description: Creates a booking with the given payload
      parameters:
      - description: Booking Payload
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/domain.Booking'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new booking
      tags:
      - booking
  /health:
    get:
      description: Checks if the database connection is alive
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check service health
      tags:
      - health
  /summary/vendor/{vendor_id}:
    get:
      consumes:
      - application/json
      description: Returns booking summary for a given vendor
      parameters:
      - description: Vendor ID
        in: path
        name: vendor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.VendorSummary'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get summary for a vendor
      tags:
      - Vendor
swagger: "2.0"